refactor: separate ball to own file
